refactor(factor): move steering factor interface to motion utils (#9337)
